+++ ChestsAnywhere/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added Automate options to block input or output for a chest.
+* Added owner's name to cabin categories.
 * Separated categories for locations with the same name.
 * Improved translations. Thanks to kelvindules (updated Portuguese) and kice (updated Chinese)!